Move find_sevctl_or_exit to utils + some fixes & cleanup
